pull most randomizing in (Abstract)RomHandler into FooRandomizer classes
The structure, with randomizers and services in their
own packages, is significantly inspired by IronHideElvan's
branch, and in some cases is intentionally similar.
Having used it as reference, some credit must of course go to them.

Since the methods are more or less untouched,
except for the move, this should be reasonably safe.
I don't feel the need to write test cases to
ensure it went well.

The Randomizer class has yet to be looked over,
nor have existing test cases, so the program does not compile
nor work in this state.